cert/pyproject.toml
[tool.poetry]
name = "datarightsprotocol"
version = "0.4.0"
description = "datarightsprotocol.org"
authors = ["Ryan Rix <ryan.rix.consultant@consumer.org>"]
license = "Apache 2"

[tool.poetry.dependencies]
python = "^3.9"
pyjwt = "^2.3.0"
click = "^8.0.3"
fastapi = "^0.73.0"
uvicorn = "^0.17.4"
pytest = "^7.0.0"
swagger-ui-py = "^21.12.8"

[tool.poetry.dependencies.pydantic]
extras = ["email"]
version = "^1.9.0"

[tool.poetry.scripts]
genjwts = 'datarightsprotocol.tools.genjwts:cmd'
genreqs = 'datarightsprotocol.tools.genreqs:cmd'
statusserver = 'datarightsprotocol.tools.status_server:start'
swagger = 'datarightsprotocol.tools.swagger:start'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

